# docs/api.yaml
openapi: 3.0.3
info:
  title: CareConnect API
  description: |
    API for managing medication prescriptions and sending notifications to elderly users.
    This system helps users and caretakers manage medication schedules through timely notifications.
  version: 1.0.0
  contact:
    name: API Support
    email: support@mednotify.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://careconnect.lmezencio.dev/api/v1
    description: Production server
...
security:
  - FirebaseJWT: []
paths:
...
  /users/me/device-tokens:
    post:
      tags:
        - DeviceTokens
      summary: Register a device token
      description: |
        Register or re-register an FCM device token for the authenticated user.
        Idempotent for the same (user, token) pair; returns 409 if the token
        already belongs to a different user.
      operationId: registerDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceTokenRequest'
      responses:
        '201':
          description: Token registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags:
        - DeviceTokens
      summary: List the caller's device tokens
      operationId: listDeviceTokens
      responses:
        '200':
          description: Device tokens for the authenticated user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceToken'
  /users/me/device-tokens/{tokenId}:
    parameters:
      - name: tokenId
        in: path
        required: true
        schema:
          type: string
    delete:
      tags:
        - DeviceTokens
      summary: Delete a device token
      operationId: deleteDeviceToken
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - DeviceTokens
      summary: Enable or disable a device token
      operationId: setDeviceTokenEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDeviceTokenEnabledRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # LGPD data-export endpoint is implemented as an extended route outside
  # the generated OpenAPI server (see internal/api/router.go). The response
  # schema (UserDataExport) is still defined below for documentation.

  # Doctor endpoints
...
components:
  securitySchemes:
    FirebaseJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase JWT token for authentication. The `Authorization` header
        accepts either the standard form `Bearer <token>` (scheme
        case-insensitive) or the bare Firebase ID token returned by
        `/auth/login` — both are accepted so clients can paste the token
        straight from the login response without adding a prefix.
        Non-Bearer schemes (e.g. `Basic`, `Token`) and malformed tokens
        are rejected with `401 invalid authorization format`.
      x-issuer: Firebase

    DemoSecret:
      type: apiKey
      name: Authorization
      in: header
      description: Demo secret for prescription creation
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
        details:
          type: string
          description: Additional error details
          example: "Field 'name' is required"
      required:
        - error

    # Auth schemas
...
    DeviceToken:
      type: object
      required: [id, user_id, enabled, created_at, updated_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
    RegisterDeviceTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 4
    SetDeviceTokenEnabledRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request"
            details: "Field 'name' is required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not found"
            details: "User with ID '123' not found"
    Forbidden:
      description: Forbidden - the caller is not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "access denied"
            details: "Token belongs to another user"
    Conflict:
      description: Resource already exists or belongs to another user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "conflict"
            details: "Token is already registered to a different user"
...